[FIX] product: pricelist item deterministic order
Following commit a8c99815c13a918af89e1b960f1b6cbdb02e755a

The order of the pricelist items on the interface was made deterministic, but
there was still a problem with the actual computation because the SQL was
created manually and used a different order than the _order of the model.

Moreover, it is better to use id DESC instead of ASC to make sure the
automatically created default rule matching everything is actually applied last.

PR: #28476