[FIX] odoo: effectively allow to use 'use' in value xml tag
Purpose is to allow to use the 'use' tag allowed by _eval_xml (see convert.py
for more details). It allows to read a specific field instead of id when using
a search tag in xml. Feature is already supported; this commit just enables
it in order to use it in demo data notably.

This commit is linked to task ID 1907994 and PR #28533.