[FIX] mail: mail queue is interrupted if a recipient is not ascii valid
This is related to revision
3d94922f5d7e451b0b5e148688bc07e1df35257e

This is nice to have a meaningful explanation
for the user to understand why his mail is not sent,

this is better to not interrupt the mail queue when doing so.

Because of the above revision,
when the mail queue iterated on a mail failing because
of an ASCII encoding issue,
the mail queue was interrupted,
and on the next cron call it failed again and again
on the same mail, therefore leading to the mail queue
to never be processed.

Fixes #27804