[FIX] website_event: restore event list layout
Commit https://github.com/odoo/odoo/commit/e981645fb9a08b1aa1d369d71b217abd03ab7e90
introduced a row in a row. On a standard database, the bug is invisible.
When the user wants to edit that event list page, the right column... is
not on the right anymore.

Note: in 12.0, thanks to BS4, the bug was revealed even more as even on
a standard non-edited database, the bug is visible.

Closes https://github.com/odoo/odoo/pull/28693