[FIX] mail: execute only once when setting automated actions at record update
Fine-tuning of commit: d470d1535b105469433ade57996eec5e1f6a149b
Which applied when the next action was to create a new activity.
However the problem is also present when the next action is to send an email.
We extract the logic in a new private function to share the code in both cases.

opw 1904156

closes odoo/odoo#28819